Mon, Jul 31, 2023, 11:55 AM +04:30

+++ CS50x/PSET2/readability/readability.c
@@ -11,9 +11,9 @@ int main(void)
 
     //Calculating the letters, words and sentences
 
-    int numC = 0;
-    int numW = 1;
-    int numS = 0;
+    float numC = 0;
+    float numW = 1;
+    float numS = 0;
 
     for(int i =0; i < strlen(text); i++)
     {
@@ -42,11 +42,11 @@ int main(void)
 
     //Calculating the grade
 
-    int L = numC*100/numW;
-    int S = numS*100/numW;
+    float L = numC*100/numW;
+    float S = numS*100/numW;
     float index = 0.0588 * L - 0.296 * S - 15.8;
     printf("%f\n", index);
-    
+
     int rounded = round(index);
 
     //Printing the grade
